+++ shuup/xtheme/admin_module/views.py
@@ -6,16 +6,20 @@
 # LICENSE file in the root directory of this source tree.
 from __future__ import unicode_literals
 
+import warnings
+
 from django import forms
 from django.contrib import messages
 from django.core.urlresolvers import reverse
 from django.http.response import HttpResponseRedirect
 from django.utils.translation import ugettext_lazy as _
+from django.views.generic import TemplateView
 from django.views.generic.edit import FormView
 
 from shuup.admin.utils.views import CreateOrUpdateView
 from shuup.admin.views.wizard import TemplatedWizardFormDef, WizardPane
 from shuup.apps.provides import get_provide_objects
+from shuup.utils.deprecation import RemovedInFutureShuupWarning
 from shuup.xtheme._theme import (
     get_current_theme, get_theme_by_identifier, set_current_theme
 )
@@ -113,7 +117,14 @@ class ThemeConfigDetailView(CreateOrUpdateView):
 
     def get_context_data(self, **kwargs):
         context = super(ThemeConfigDetailView, self).get_context_data(**kwargs)
-        context["theme"] = self.get_theme()
+        context["theme"] = theme = self.get_theme()
+        context["active_stylesheet"] = self.object.data.get("settings", {}).get("stylesheet", None)
+        if isinstance(theme.stylesheets[0], dict):
+            context["has_images"] = True
+        else:
+            warnings.warn(
+                "Using list of tuples in theme.stylesheets will deprecate "
+                "in Shuup 0.5.7. Use list of dictionaries instead.", RemovedInFutureShuupWarning)
         return context
 
     def get_form(self, form_class=None):
@@ -123,3 +134,12 @@ class ThemeConfigDetailView(CreateOrUpdateView):
         return reverse("shuup_admin:xtheme.config_detail", kwargs={
             "theme_identifier": self.object.theme_identifier
         })
+
+
+class ThemeGuideTemplateView(TemplateView):
+    template_name = None
+
+    def dispatch(self, request, *args, **kwargs):
+        theme = get_theme_by_identifier(kwargs["theme_identifier"])
+        self.template_name = theme.guide_template
+        return super(ThemeGuideTemplateView, self).dispatch(request, *args, **kwargs)